pyproject.toml
[tool.poetry]
name = "playlist"
version = "0.1.0"
description = "vivpro take home assignment"
authors = ["Anshuman"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
